[241002] 김민정 Day25 필수과제 제출
