+++ 메인캠프/maincamp/my-board/src/app/boards/new/styles.module.css
@@ -7,6 +7,10 @@
   flex-direction: column;
   align-items: flex-start;
 }
+.inputCSS:disabled {
+  background-color: #e2e2e2;
+  cursor: not-allowed;
+}
 .textareaCSS {
   border-radius: 0.5rem;
   border: 1px solid var(--Gray-Gray-200, #d4d3d3);
